CFO_AGENTS/CashFlow_Agents/config.py
import os
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
load_dotenv()

GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

llm = ChatGoogleGenerativeAI(
    model="gemini-2.0-flash",   # Gemini 3.0 equivalent
    temperature=0.2,
    api_key=GOOGLE_API_KEY
)
